added ads list screen

+++ src/api/AxiosInceptor.js
@@ -1,9 +1,26 @@
 import axios from "axios";
+import { ApiUrl } from "./apiUrls";
+
+let isRefreshing = false;
+let failedQueue = [];
+
+const processQueue = (error, token = null) => {
+  failedQueue.forEach((prom) => {
+    if (error) {
+      prom.reject(error);
+    } else {
+      prom.resolve(token);
+    }
+  });
+  failedQueue = [];
+};
+
 const api = axios.create({
   baseURL:
     `${process.env.REACT_APP_API_URL}/api` || "http://localhost:5000/api",
   timeout: 20000,
 });
+
 api.interceptors.request.use(
   (config) => {
     const token = localStorage.getItem("Token");
@@ -14,14 +31,52 @@ api.interceptors.request.use(
   },
   (error) => Promise.reject(error),
 );
+
 api.interceptors.response.use(
   (response) => response,
-  (error) => {
-    if (error.response?.status === 401) {
-      localStorage.clear();
-      // window.location.href = "/login";
+  async (error) => {
+    const originalRequest = error.config;
+
+    if (error.response?.status === 401 && !originalRequest._retry) {
+      if (isRefreshing) {
+        return new Promise((resolve, reject) => {
+          failedQueue.push({ resolve, reject });
+        })
+          .then((token) => {
+            originalRequest.headers.Authorization = `Bearer ${token}`;
+            return api(originalRequest);
+          })
+          .catch((err) => Promise.reject(err));
+      }
+
+      originalRequest._retry = true;
+      isRefreshing = true;
+
+      try {
+        const user = localStorage.getItem("User");
+        const response = await axios.post(ApiUrl.Auth.refreshToToken, {
+          refreshToken: user?.refreshToken,
+        });
+
+        const { token: newAccessToken } = response.data;
+        localStorage.setItem("Token", newAccessToken);
+
+        processQueue(null, newAccessToken);
+
+        originalRequest.headers.Authorization = `Bearer ${newAccessToken}`;
+        return api(originalRequest);
+      } catch (refreshError) {
+        processQueue(refreshError, null);
+        localStorage.clear();
+        window.location.href = "/login";
+        return Promise.reject(refreshError);
+      } finally {
+        isRefreshing = false;
+      }
     }
+
     return Promise.reject(error);
   },
 );
+
 export default api;
